Move Invocation into own file
