+++ SimplyAOP/IAspect.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Reflection;
 
 namespace SimplyAOP
 {
@@ -22,23 +21,5 @@ namespace SimplyAOP
         void AfterReturning<TResult>(Invocation invocation, ref TResult result);
 
         void AfterThrowing(Invocation invocation, ref Exception exception);
-    }
-
-    public struct Invocation
-    {
-        private readonly Lazy<Type> targetType;
-        private readonly Lazy<MethodInfo> method;
-
-        public Invocation(Lazy<Type> targetType, string methodName)
-        {
-            MethodName = methodName;
-            this.targetType = targetType;
-
-            method = new Lazy<MethodInfo>(() => targetType.Value.GetMethod(methodName));
-        }
-
-        public Type TargetType => targetType.Value;
-
-        public string MethodName { get; }
     }
 }
